Move device detection code to new module
